add seperate files for the logger to clean up the code
